backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-waffle==4.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dotenv==1.1.0
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
